requirements.txt
aiogram==2.25.1
python-dotenv==0.19.2
aiohttp==3.8.4
Pillow==9.5.0
SQLAlchemy==1.4.41
psycopg2-binary==2.9.5
alembic==1.12.0
python-json-logger==2.0.7
prometheus-client==0.17.1
redis==4.6.0
